+++ packages/profile-header/index.test.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
-import { mount } from 'enzyme';
+import { configure, mount } from 'enzyme';
+import Adapter from 'enzyme-adapter-react-16';
 import { Provider } from 'react-redux';
 import Header, {
   reducer,
@@ -9,6 +10,7 @@ import Header, {
 } from './index';
 import ProfileHeader from './components/ProfileHeader';
 
+configure({ adapter: new Adapter() });
 const storeFake = state => ({
   default: () => {},
   subscribe: () => {},
