+++ packages/profile-selector/index.test.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
-import { shallow } from 'enzyme';
+import { configure, shallow } from 'enzyme';
+import Adapter from 'enzyme-adapter-react-16';
 import configureMockStore from 'redux-mock-store';
 
 import ProfileSelector, {
@@ -8,6 +9,7 @@ import ProfileSelector, {
   actionTypes,
 } from './index';
 
+configure({ adapter: new Adapter() });
 describe('ProfileSelector', () => {
   let profiles = [];
   let state = {};
